requirements.txt
pytest==8.3.2
Flask==3.0.3
Flask-Limiter==3.8.0
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1